+++ color/term.go
@@ -1,0 +1,5 @@
+package color
+
+func Term(text, colorConst string) string {
+	return colorConst + text + "\033[0m"
+}