+++ text/bold.go
@@ -1,0 +1,5 @@
+package text
+
+func Bold(prompt string) string {
+	return "\033[1m" + prompt + "\033[0m"
+}