+++ text/cursive.go
@@ -1,0 +1,5 @@
+package text
+
+func Cursive(prompt string) string {
+	return "\033[3m" + prompt + "\033[0m"
+}