+++ text/strike.go
@@ -1,0 +1,5 @@
+package text
+
+func Strike(prompt string) string {
+	return "\033[9m" + prompt + "\033[0m"
+}