+++ text/underline.go
@@ -1,0 +1,5 @@
+package text
+
+func Underline(prompt string) string {
+	return "\033[4m" + prompt + "\033[0m"
+}